+++ Editor/Drawers/Toolbox/PropertySelf/DynamicMinMaxSliderAttributeDrawer.cs
@@ -1,20 +1,49 @@
-using UnityEditor;
+﻿using UnityEditor;
 using UnityEngine;
 
 namespace Toolbox.Editor.Drawers
 {
+    using Toolbox.Editor.Internal;
+
     public class DynamicMinMaxSliderAttributeDrawer : DynamicMinMaxBaseDrawer<DynamicMinMaxSliderAttribute>
     {
         protected override void OnGuiSafe(SerializedProperty property, GUIContent label, float minValue, float maxValue)
         {
-            var xValue = property.vector2Value.x;
-            var yValue = property.vector2Value.y;
+            var xValue = 0.0f;
+            var yValue = 0.0f;
+            switch (property.propertyType)
+            {
+                case SerializedPropertyType.Vector2:
+                    xValue = property.vector2Value.x;
+                    yValue = property.vector2Value.y;
+                    break;
+                case SerializedPropertyType.Vector2Int:
+                    xValue = property.vector2IntValue.x;
+                    yValue = property.vector2IntValue.y;
+                    break;
+            }
+
             ToolboxEditorGui.BeginProperty(property, ref label, out var position);
+            position = EditorGUI.PrefixLabel(position, label);
             EditorGUI.BeginChangeCheck();
-            ToolboxEditorGui.DrawMinMaxSlider(position, label, ref xValue, ref yValue, minValue, maxValue);
+            using (new ZeroIndentScope())
+            {
+                ToolboxEditorGui.DrawMinMaxSlider(position, ref xValue, ref yValue, minValue, maxValue);
+            }
+
             if (EditorGUI.EndChangeCheck())
             {
-                property.vector2Value = new Vector2(xValue, yValue);
+                switch (property.propertyType)
+                {
+                    case SerializedPropertyType.Vector2:
+                        property.vector2Value = new Vector2(xValue, yValue);
+                        break;
+                    case SerializedPropertyType.Vector2Int:
+                        var intXValue = Mathf.RoundToInt(xValue);
+                        var intYValue = Mathf.RoundToInt(yValue);
+                        property.vector2IntValue = new Vector2Int(intXValue, intYValue);
+                        break;
+                }
             }
 
             ToolboxEditorGui.CloseProperty();
@@ -23,7 +52,8 @@ namespace Toolbox.Editor.Drawers
 
         public override bool IsPropertyValid(SerializedProperty property)
         {
-            return property.propertyType == SerializedPropertyType.Vector2;
+            return property.propertyType == SerializedPropertyType.Vector2 || 
+                property.propertyType == SerializedPropertyType.Vector2Int;
         }
     }
 }