+++ Editor/IToolboxEditor.cs
@@ -10,7 +10,13 @@ namespace Toolbox.Editor
         void IgnoreProperty(SerializedProperty property);
         void IgnoreProperty(string propertyPath);
 
+        /// <summary>
+        /// <see cref="Editor"/> instance associated to this Editor.
+        /// </summary>
         Editor ContextEditor { get; }
+        /// <summary>
+        /// Dedicated "drawer" that is responsible for default drawing strategy for this <see cref="ToolboxEditor"/>.
+        /// </summary>
         IToolboxEditorDrawer Drawer { get; }
     }
 }