+++ Editor/ToolboxEditor.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-
 using UnityEditor;
 
 namespace Toolbox.Editor
@@ -45,8 +44,8 @@ namespace Toolbox.Editor
             Drawer.IgnoreProperty(propertyPath);
         }
 
-
         Editor IToolboxEditor.ContextEditor => this;
+        /// <inheritdoc />
         public virtual IToolboxEditorDrawer Drawer { get; } = new ToolboxEditorDrawer();
 
 #pragma warning disable 0067