+++ Editor/ToolboxEditorDrawer.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-
 using UnityEditor;
 
 namespace Toolbox.Editor
@@ -15,7 +14,12 @@ namespace Toolbox.Editor
         private readonly Action<SerializedProperty> toolboxDrawingAction;
         private readonly Action<SerializedProperty> defaultDrawingAction;
 
-        public ToolboxEditorDrawer() : this(ToolboxEditorGui.DrawToolboxProperty, ToolboxEditorGui.DrawNativeProperty)
+        public ToolboxEditorDrawer()
+            : this(ToolboxEditorGui.DrawToolboxProperty)
+        { }
+
+        public ToolboxEditorDrawer(Action<SerializedProperty> toolboxDrawingAction)
+            : this(toolboxDrawingAction, ToolboxEditorGui.DrawNativeProperty)
         { }
 
         public ToolboxEditorDrawer(Action<SerializedProperty> toolboxDrawingAction, Action<SerializedProperty> defaultDrawingAction)
@@ -26,12 +30,10 @@ namespace Toolbox.Editor
 
         private void DrawProperty(SerializedProperty property, Action<SerializedProperty> drawingAction)
         {
-            if (IsPropertyIgnored(property))
+            if (!IsPropertyIgnored(property))
             {
-                return;
+                drawingAction(property);
             }
-
-            drawingAction(property);
         }
 
         /// <inheritdoc />