+++ Editor/ToolboxPropertyHandler.cs
@@ -7,6 +7,7 @@ using UnityEngine;
 
 namespace Toolbox.Editor
 {
+    using Toolbox.Attributes.Property;
     using Toolbox.Editor.Drawers;
 
     /// <summary>
@@ -83,6 +84,11 @@ namespace Toolbox.Editor
         /// </summary>
         private bool hasToolboxConditionDrawer;
 
+        /// <summary>
+        /// Temporary solution to cache all attributes responsible for overriding label.
+        /// This will be replaced with validation attributes.
+        /// </summary>
+        private ILabelProcessorAttribute labelProcessorAttribute;
 
         /// <summary>
         /// Constructor prepares all property-related data for custom drawing.
@@ -166,19 +172,28 @@ namespace Toolbox.Editor
 
         private void HandleNewAttribute(PropertyAttribute attribute)
         {
-            if (!isChild)
+            //NOTE: setting tooltip and labels is valid only for parent or single properties
+            //it's a bit ugly but, it's the only semi-acceptable way to support built-in TooltipAttribute
+            //TODO: move these attributes to validation drawers
+            switch (attribute)
             {
-                //NOTE: setting tooltip and labels is valid only for parent or single properties
-                //it's a bit ugly but, it's the only semi-acceptable way to support built-in TooltipAttribute
-                switch (attribute)
-                {
-                    case TooltipAttribute a:
-                        label.tooltip = a.tooltip;
-                        return;
-                    case NewLabelAttribute a:
-                        label.text = a.NewLabel;
-                        return;
-                }
+                case TooltipAttribute a:
+                    label.tooltip = a.tooltip;
+                    return;
+                case NewLabelAttribute a:
+                    if (!isChild)
+                    {
+                        labelProcessorAttribute = a;
+                    }
+
+                    return;
+                case LabelByChildAttribute a:
+                    if (!isArray)
+                    {
+                        labelProcessorAttribute = a;
+                    }
+
+                    return;
             }
 
             var attributeType = attribute.GetType();
@@ -269,6 +284,8 @@ namespace Toolbox.Editor
 
         private void DrawProperty(SerializedProperty property, GUIContent label)
         {
+            ProcessLabel(property, label);
+
             //get toolbox drawer for the property or draw it in the default way
             if (hasToolboxPropertyDrawer && (!hasBuiltInPropertyDrawer || isArray))
             {
@@ -360,6 +377,24 @@ namespace Toolbox.Editor
             return ToolboxDrawerModule.GetConditionDrawer(conditionAttribute)?.OnGuiValidate(property, conditionAttribute) ?? PropertyCondition.Valid;
         }
 
+        //TODO: replace this method with validation attributes
+        private void ProcessLabel(SerializedProperty property, GUIContent label)
+        {
+            if (labelProcessorAttribute == null)
+            {
+                return;
+            }
+
+            switch (labelProcessorAttribute)
+            {
+                case NewLabelAttribute a:
+                    label.text = a.NewLabel;
+                    return;
+                case LabelByChildAttribute a:
+                    PropertyUtility.OverrideLabelByChild(label, property, a.ChildName);
+                    return;
+            }
+        }
 
         /// <summary>
         /// Draw property using built-in layout system and cached <see cref="ToolboxAttributeDrawer"/>s.